make: only run test on matching version number
The tests need the exact installed version to run the correct
tests so we need to enforce it even if it mostly works otherwise.

The procedure is mostly used for the nightly builds anyway, and
that shall run directly after the build.

(cherry picked from commit d5a795b329fd56c39d67fabb20816d22b4160a05)